refactor: update nodes structures
