--- frontend/src/app/features/WorkflowNodes/CustomNodes.tsx
@@ -1,4 +1,0 @@
-export { NodeCreate } from './NodeCreate';
-export { NodeWebhook } from './NodeWebhook';
-export { NodeHttp } from './NodeHttp';
-export { NodeDiscord } from './NodeDiscord';
