+++ frontend/src/app/features/WorkflowNodes/NodeHttp.tsx
@@ -1,7 +1,7 @@
 import { Globe } from 'lucide-react';
 import { NodeBase } from './NodeBase';
 import { memo } from 'react';
-import { NodeConfigPopover } from './NodeConfigPopover';
+import { NodeConfigPopover } from '../WorkflowConfig/NodeConfigPopover';
 
 interface NodeData {
   // Add HTTP specific data fields here
